requirements/base.txt
django==3.1.7
requests==2.25.1
psycopg2==2.9.1
Pillow==8.2.0
celery==5.0.5
amqp==5.0.6
django-celery-results==2.0.1
djangorestframework==3.12.4
